requirements.txt
Django
django-allauth
markdown
django-filter
pillow
djangorestframework
django-cart
django-crispy-forms
stripe
django-utils-six
django_countries
django-allauth
django-rest-auth
dj-rest-auth
django-decouple
django-comments-dab==1.6.7
django-widget-tweaks
django-bootstrap4
django-star-ratings
az-iranian-bank-gateways


